// src/types/Birds.ts
export interface Bird {
    id: string;
    name: string;
    scientificName: string;
    otherNames: string[];
    description: string;
    points: number;
    rarity: string;
    activity: string;
    diet: string[];
    habitat: string[];
    movement: string;
    conservationStatus: string;
    imageUrl: string;
}